feat: タスク 23.4 handlers.py の分割
- handlers.py (1,584行) を cell_handlers.py / contents_handlers.py /
  preview_handlers.py に分割
- 共有ヘルパー (validate_path, _apply_lock_token) を base.py に集約
- _resolve_workspace_for_kernel を resolve_workspace_for_kernel にリネーム
  (モジュール境界を越える public API のため _ prefix を除去)
- テストの base モック に validate_path / _apply_lock_token を追加
