+++ docs/plan/01-jupyter.md
@@ -27,7 +27,7 @@
 | 23.1 | カーネル単位の実行直列化 | [x] | 同一カーネルへ 2 つのコード実行を並行リクエストしても、各レスポンスに他方の出力が混入しない（並行・異常系） | I6 違反の解消。`kernel_executor.py` の実行をカーネル単位で直列化 |
 | 23.2 | 同期 I/O のイベントループ外オフロード | [x] | 大きな CSV のプレビュー処理中も他の API リクエストが応答する（並行・異常系） | I3 違反の解消。`handlers.py` の preview・`workspace_handlers.py` の同期 I/O を run_in_executor へ |
 | 23.3 | SQL エンジンのプール化とタイムアウト後の接続解放 | [x] | 連続 SQL 実行で DB 接続数が増え続けない。クエリタイムアウト後も後続クエリが成功する（タイムアウト・異常系） | リクエストごとの create_engine + dispose を廃止しプール共有。wait_for 残置スレッドによるプール枯渇対策を含む |
-| 23.4 | handlers.py の分割 | [ ] | 既存テストが全件成功し、分割後の各ハンドラファイルが肥大していない（行数検証コマンドは詳細計画で確定） | 1,515 行のモノリスを cell_actions / contents / preview に分割。挙動を変えない純リファクタ |
+| 23.4 | handlers.py の分割 | [→] | 既存テストが全件成功し、分割後の各ハンドラファイルが肥大していない（行数検証コマンドは詳細計画で確定） | 1,515 行のモノリスを cell_actions / contents / preview に分割。挙動を変えない純リファクタ |
 | 23.5 | sandbox の os ホワイトリスト統一 | [ ] | `os.rename` / `shutil.move` によるワークスペース外へのファイル移動が拒否される（不正入力・異常系） | `workspace_sandbox.py` / `code_validator.py` の os 関数をブロックリストからホワイトリスト方式へ統一 |
 | 23.6 | 軽微な負債の解消 | [ ] | DATABASE_URL 未設定時のエラーステータスが実行系・エクスポート系で一致する | §4-6/7: except 握りつぶし・Contents POST 重複・タイムアウトエラー組み立て3重複・`sync_state.py` `_seq_store` 残置エントリの掃除 |
