+++ jupyter-server/tests/test_cell_merge_split.py
@@ -1,6 +1,6 @@
 """ContentsCellsHandler の merge/split アクションのユニットテスト
 
-handlers.py の ContentsCellsHandler.patch() に merge/split アクションを追加する。
+cell_handlers.py の ContentsCellsHandler.patch() に merge/split アクションを追加する。
 ハンドラーの処理ロジックを直接インポートしてモックでテストする。
 
 テスト対象のアクションがまだ存在しないため（TDD Red フェーズ）、
@@ -46,15 +46,9 @@ if "custom_api.base" not in sys.modules:
     _base_mock.WORKSPACE_PATH_PREFIX = "workspaces/sample"
     _base_mock.JUPYTER_ROOT_DIR = "/home/jovyan/work"
     _base_mock.validate_kernel_name = lambda *a, **kw: None
+    _base_mock.validate_path = lambda path, *a, **kw: path
+    _base_mock._apply_lock_token = lambda handler: None
     sys.modules["custom_api.base"] = _base_mock
-
-# ai_events モジュールのモック
-if "custom_api.ai_events" not in sys.modules:
-    _ai_events_mock = _types.ModuleType("custom_api.ai_events")
-    _ai_events_mock.__package__ = "custom_api"
-    _ai_events_mock.AiEventsWebSocketHandler = type("AiEventsWebSocketHandler", (), {})
-    _ai_events_mock.AiEventsPostHandler = type("AiEventsPostHandler", (), {})
-    sys.modules["custom_api.ai_events"] = _ai_events_mock
 
 # code_validator モジュール（実際にロード — 純粋関数のため）
 if "custom_api.code_validator" not in sys.modules:
@@ -74,46 +68,20 @@ if "custom_api.kernel_executor" not in sys.modules:
     _ke_mock.KernelExecutor = MagicMock
     sys.modules["custom_api.kernel_executor"] = _ke_mock
 
-# session_handlers モジュールのモック
-if "custom_api.session_handlers" not in sys.modules:
-    _sh_mock = _types.ModuleType("custom_api.session_handlers")
-    _sh_mock.__package__ = "custom_api"
-    _sh_mock.CustomSessionsHandler = type("CustomSessionsHandler", (), {})
-    _sh_mock.get_kernel_workspace = lambda *a: None
-    _sh_mock.unregister_kernel = lambda *a: None
-    sys.modules["custom_api.session_handlers"] = _sh_mock
-
-# sql_handlers モジュールのモック
-if "custom_api.sql_handlers" not in sys.modules:
-    _sql_mock = _types.ModuleType("custom_api.sql_handlers")
-    _sql_mock.__package__ = "custom_api"
-    _sql_mock.SqlExecuteHandler = type("SqlExecuteHandler", (), {})
-    _sql_mock.SqlExportHandler = type("SqlExportHandler", (), {})
-    sys.modules["custom_api.sql_handlers"] = _sql_mock
-
-# workspace_handlers モジュールのモック
-if "custom_api.workspace_handlers" not in sys.modules:
-    _wh_mock = _types.ModuleType("custom_api.workspace_handlers")
-    _wh_mock.__package__ = "custom_api"
-    _wh_mock.WorkspaceHandler = type("WorkspaceHandler", (), {})
-    _wh_mock.WorkspacesHandler = type("WorkspacesHandler", (), {})
-    _wh_mock.WorkspaceSummarizeHandler = type("WorkspaceSummarizeHandler", (), {})
-    sys.modules["custom_api.workspace_handlers"] = _wh_mock
-
-# --- 3. handlers モジュールをロード ---
-_module_path = _ext_dir / "custom_api" / "handlers.py"
-_handlers_spec = importlib.util.spec_from_file_location(
-    "custom_api.handlers",
+# --- 3. cell_handlers モジュールをロード ---
+_module_path = _ext_dir / "custom_api" / "cell_handlers.py"
+_cell_handlers_spec = importlib.util.spec_from_file_location(
+    "custom_api.cell_handlers",
     _module_path,
     submodule_search_locations=[],
 )
-_handlers = importlib.util.module_from_spec(_handlers_spec)
-_handlers.__package__ = "custom_api"
-sys.modules["custom_api.handlers"] = _handlers
-_handlers_spec.loader.exec_module(_handlers)
+_cell_handlers = importlib.util.module_from_spec(_cell_handlers_spec)
+_cell_handlers.__package__ = "custom_api"
+sys.modules["custom_api.cell_handlers"] = _cell_handlers
+_cell_handlers_spec.loader.exec_module(_cell_handlers)
 
-validate_path = _handlers.validate_path
-ContentsCellsHandler = _handlers.ContentsCellsHandler
+validate_path = _cell_handlers.validate_path
+ContentsCellsHandler = _cell_handlers.ContentsCellsHandler
 
 
 # ============================================================
